centralised time-util functions and removed pointless async
